+++ src/main/java/groupwork/dto/SingerDTO.java
@@ -4,9 +4,9 @@ import java.util.Objects;
 
 public class SingerDTO {
     private String name;
-    private int id;
+    private long id;
 
-    public SingerDTO(String name, int id) {
+    public SingerDTO(String name, long id) {
         this.name = name;
         this.id = id;
     }
@@ -26,11 +26,11 @@ public class SingerDTO {
         this.name = name;
     }
 
-    public int getId() {
+    public long getId() {
         return id;
     }
 
-    public void setId(int id) {
+    public void setId(long id) {
         this.id = id;
     }
